Moves group permissions decoder to helper
