+++ src/app/shared/http-services/group-permissions.service.ts
@@ -7,19 +7,7 @@ import { assertSuccess, SimpleActionResponse } from './action-response';
 import * as D from 'io-ts/Decoder';
 import { pipe } from 'fp-ts/function';
 import { decodeSnakeCase } from 'src/app/shared/operators/decode';
-
-export const permissionsDecoder = pipe(
-  D.struct({
-    canView: D.literal('none','info','content','content_with_descendants','solution'),
-    canWatch: D.literal('none','result','answer','answer_with_grant'),
-    canEdit: D.literal('none','children','all','all_with_grant'),
-    canGrantView: D.literal('none','enter','content','content_with_descendants','solution','solution_with_grant'),
-    isOwner: D.boolean,
-    canMakeSessionOfficial: D.boolean,
-  })
-);
-
-export type Permissions = D.TypeOf<typeof permissionsDecoder>;
+import { permissionsDecoder, Permissions } from '../helpers/group-permissions';
 
 const groupPermissionsDecoder = pipe(
   D.struct({
